feat: form widget in a component
