+++ src/components/index.js
@@ -1,5 +1,6 @@
 import DeleteWidget from './DeleteWidget';
 import FabWidget from './FabWidget';
+import FormWidget from './FormWidget';
 import MultiInput from './MultiInput';
 import Navbar from './Navbar';
 import WidgetCard from './WidgetCard';
@@ -7,6 +8,7 @@ import WidgetCard from './WidgetCard';
 export {
   DeleteWidget,
   FabWidget,
+  FormWidget,
   MultiInput,
   Navbar,
   WidgetCard,
